refactor(ui): extract ViewportTermControls from Viewport (#855)
Move the mobile ctrl-row + compose unit (ControlBars, notes/attach/
dictate/enter chips, ComposeBar, swipe-up summon $effect, file input)
+ scoped CSS into a child. conn passed as a send() callback; shared
attachImages/notesKey/enter/uploading stay in Viewport as props.
Viewport <template> 120/180 -> 107/165; child 11/15.
